+++ src/Semantic_Enum.cpp
@@ -345,7 +345,7 @@ bool Semantic::resolveEnumValue(SemanticContext* context)
 
         // First child is enumType
         const AstNode* firstEnumValue = nullptr;
-        for (size_t fev = 1; fev < enumNode->children.size(); fev++)
+        for (uint32_t fev = 1; fev < enumNode->children.size(); fev++)
         {
             firstEnumValue = enumNode->children[fev];
             if (firstEnumValue->kind == AstNodeKind::EnumValue && !firstEnumValue->hasSpecFlag(AstEnumValue::SPEC_FLAG_HAS_USING))
