+++ src/Semantic_Function.cpp
@@ -441,7 +441,7 @@ void Semantic::setFuncDeclParamsIndex(const AstFuncDecl* funcNode)
         }
 
         const auto childSize = funcNode->parameters->children.size();
-        for (size_t i = 0; i < childSize; i++)
+        for (uint32_t i = 0; i < childSize; i++)
         {
             if (i == childSize - 1 && funcNode->typeInfo->hasFlag(TYPEINFO_VARIADIC | TYPEINFO_TYPED_VARIADIC))
                 break;
